+++ src/org/opensha2/eq/model/PointSourceFixedStrike.java
@@ -7,7 +7,7 @@ import static java.lang.Math.sqrt;
 import static java.lang.Math.tan;
 
 import static org.opensha2.eq.fault.FocalMech.STRIKE_SLIP;
-import static org.opensha2.geo.GeoTools.TO_RAD;
+import static org.opensha2.geo.Coordinates.TO_RAD;
 import static org.opensha2.util.Maths.hypot;
 
 import org.opensha2.data.XySequence;
