+++ src/org/opensha2/geo/LocationVector.java
@@ -5,8 +5,8 @@ import static java.lang.Math.atan;
 import static java.lang.Math.cos;
 import static java.lang.Math.sin;
 
-import static org.opensha2.geo.GeoTools.TO_DEG;
-import static org.opensha2.geo.GeoTools.TWOPI;
+import static org.opensha2.geo.Coordinates.TO_DEG;
+import static org.opensha2.geo.Coordinates.TWOPI;
 import static org.opensha2.geo.Locations.azimuthRad;
 import static org.opensha2.geo.Locations.horzDistance;
 import static org.opensha2.geo.Locations.vertDistance;
